Reorganize the project to work with Nginx and Express
Rename the build folder to public for clarity, move server.js to the
root folder (required by Express), add parent folder to file links in
HTML files (required by Express), and make several small changes to
server.js including adding `app.set('trust proxy', true);` so we can get
users' IP address and
`app.use('/public', express.static(path.join(__dirname, "public")));`
for serving static files in the public folder.
