+++ public/about.html
@@ -5,17 +5,17 @@ Programmer Name: Jong Won Lim
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>About Us</title>
-  <link rel ="stylesheet" href="css/about.css"></link>
+  <link rel ="stylesheet" href="/public/css/about.css"></link>
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.3/css/all.css">
 </head>
 <body>
   <header>
     <nav>
       <ul>
-        <li><a href="index.html">Home</a></li>
-        <li><a href="about.html">About</a></li>
-        <li><a href="upgrade.html">Pricing</a></li>
-        <li><a href="login.html">Log In</a></li>
+        <li><a href="/public/index.html">Home</a></li>
+        <li><a href="/public/about.html">About</a></li>
+        <li><a href="/public/upgrade.html">Pricing</a></li>
+        <li><a href="/public/login.html">Log In</a></li>
       </ul>
     </nav>
   </header>
