+++ public/css/index.css
@@ -1,12 +1,13 @@
 /*
 Programmer Name: Miroo Yum
 */
-*{
+* {
     margin: 0;
     padding: 0;
     box-sizing: border-box;
     font-family: 'Poppins', sans-serif;
 }
+
 .container {
     flex-direction: column;
     align-items: center;
@@ -18,28 +19,29 @@ Programmer Name: Miroo Yum
     padding: 5%;
     display: flex;
     justify-content: center;
-  }
-.header{
+}
+
+.header {
     background: rgb(208, 218, 236);
     padding: 10px;
     text-align: right;
 }
 
-.header button{
+.header button {
     background: transparent;
-    border-color:  rgb(0, 0, 0);
+    border-color: rgb(0, 0, 0);
     padding: 5px 10px;
     font-size: 17px;
 }
 
-.Logo{
+.Logo {
     margin-bottom: 50px;
 }
 
-.searchbar{
+.searchbar {
     width: 80%;
     max-width: 700px;
-    background:mediumpurple ;
+    background: mediumpurple;
     display: flex;
     align-items: center;
     border-radius: 60px;
@@ -47,7 +49,7 @@ Programmer Name: Miroo Yum
     backdrop-filter: blur(4px) saturate(180%);
 }
 
-.searchbar input{
+.searchbar input {
     background: transparent;
     flex: 1;
     border: 0;
@@ -56,16 +58,18 @@ Programmer Name: Miroo Yum
     font-size: 20px;
     color: thistle;
 }
-::placeholder{
+
+::placeholder {
     color: white;
 }
 
-.searchbar button img{
+.searchbar button img {
     width: 60px;
     border-radius: 50%;
 
 }
-.searchbar button{
+
+.searchbar button {
     border: 0;
     border-radius: 50%;
     width: 60px;
