+++ public/index.html
@@ -1,0 +1,34 @@
+<!--
+Programmer Name: Miroo Yum
+-->
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>MakeMeShort Homepage</title>
+    <link rel="stylesheet" type="text/css" href="/public/css/index.css">
+    </link>
+  </head>
+
+  <body>
+
+    <div class="header">
+      <button id="about-btn">About</button>
+      <button id="upgrade-btn">Pricing</button>
+      <button id="login-btn">Log In</button>
+      <script src="/public/button.js"></script>
+    </div>
+
+    <div class="container">
+      <div class="Logo">
+        <span><img src="/public/images/logo.png" alt=""></span>
+      </div>
+      <form action="" class="searchbar">
+        <input type="text" placeholder="Original Link here" name="q">
+        <button type="submit"><img src="/public/images/search_icon.png" alt=""></button>
+      </form>
+    </div>
+  </body>
+
+</html>
