+++ public/login.html
@@ -3,17 +3,17 @@ Programmer Name: Jong Won Lim
 -->
 <head>
     <title>Login Page</title>
-    <link rel ="stylesheet" href="css/login.css"></link>
+    <link rel ="stylesheet" href="/public/css/login.css"></link>
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.3/css/all.css">
 </head>
 <body>
     <header>
         <nav>
           <ul>
-            <li><a href="index.html">Home</a></li>
-            <li><a href="about.html">About</a></li>
-            <li><a href="upgrade.html">Pricing</a></li>
-            <li><a href="login.html">Log In</a></li>
+            <li><a href="/public/index.html">Home</a></li>
+            <li><a href="/public/about.html">About</a></li>
+            <li><a href="/public/upgrade.html">Pricing</a></li>
+            <li><a href="/public/login.html">Log In</a></li>
           </ul>
         </nav>
       </header>
